Removed Plot container ;( and title bar
