+++ src/components/BasicPlotly/index.js
@@ -1,0 +1,2 @@
+import BasicPlotly from "./BasicPlotly.vue";
+export default BasicPlotly;
