--- src/components/PlotlyPlot/index.js
@@ -1,2 +1,0 @@
-import PlotContainer from "./PlotContainer.vue";
-export default PlotContainer;
