--- src/components/TitleBar/index.js
@@ -1,2 +1,0 @@
-import TitleBar from "./TitleBar.vue";
-export default TitleBar;
